fix model issue

+++ app/models/video.model.js
@@ -1,10 +1,10 @@
 module.exports = (sequelize, Sequelize) => {
     const Video = sequelize.define("tbl_video", {      
       user_id: {
-        type: Sequelize.BIGINT
+        type: Sequelize.INTEGER
       },
       video_id: {
-        type: Sequelize.BIGINT
+        type: Sequelize.INTEGER
       }
     });
